integer overflow in tests utility function fixed

+++ src/test/headers/s21_tests.h
@@ -17,7 +17,7 @@
 
 long double func_Uniform(double left, double right, int i) {
   srand((unsigned int)(i % MAGIC));
-  long double randomNumber = sin(rand() * rand());  // NOLINT
+  long double randomNumber = sin((double)rand() * rand());  // NOLINT
   return left + (right - left) * fabs((double)randomNumber);
 }
